refactor: refactor mcp discovery to mcp register
