+++ spring-ai-alibaba-bom/pom.xml
@@ -143,7 +143,7 @@
 
             <dependency>
                 <groupId>com.alibaba.cloud.ai</groupId>
-                <artifactId>spring-ai-alibaba-starter-nacos-mcp-discovery</artifactId>
+                <artifactId>spring-ai-alibaba-starter-nacos-mcp-register</artifactId>
                 <version>${project.version}</version>
             </dependency>
