+++ spring-ai-alibaba-mcp/spring-ai-alibaba-mcp-register/src/main/java/com/alibaba/cloud/ai/mcp/register/utils/JsonSchemaUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.cloud.ai.mcp.discovery.registry.utils;
+package com.alibaba.cloud.ai.mcp.register.utils;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
